posix: saner code organization
