+++ src/libpmemfile-posix/data.c
@@ -30,60 +30,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#define _GNU_SOURCE
-
-#include <limits.h>
-
-#include "callbacks.h"
-#include "compiler_utils.h"
+#include "block_array.h"
+#include "ctree.h"
 #include "data.h"
-#include "inode.h"
 #include "internal.h"
-#include "locks.h"
 #include "out.h"
 #include "pool.h"
-#include "os_thread.h"
 #include "valgrind_internal.h"
-#include "ctree.h"
-#include "block_array.h"
 #include "utils.h"
-
-/*
- * expand_to_full_pages
- * Alters two file offsets to be pmemfile-page aligned. This is not
- * necessarily the same as memory page alignment!
- * The resulting offset refer to an interval that contains the original
- * interval.
- */
-static void
-expand_to_full_pages(uint64_t *offset, uint64_t *length)
-{
-	/* align the offset */
-	*length += *offset % FILE_PAGE_SIZE;
-	*offset -= *offset % FILE_PAGE_SIZE;
-
-	/* align the length */
-	*length = page_roundup(*length);
-}
-
-/*
- * narrow_to_full_pages
- * Alters two file offsets to be pmemfile-page aligned. This is not
- * necessarily the same as memory page alignment!
- * The resulting offset refer to an interval that is contained by the original
- * interval. This new interval can end up being empty, i.e. *length can become
- * zero.
- */
-static void
-narrow_to_full_pages(uint64_t *offset, uint64_t *length)
-{
-	uint64_t end = page_rounddown(*offset + *length);
-	*offset = page_roundup(*offset);
-	if (end > *offset)
-		*length = end - *offset;
-	else
-		*length = 0;
-}
 
 /*
  * block_cache_insert_block -- inserts block into the tree
@@ -121,7 +75,7 @@ find_last_block(const struct pmemfile_vinode *vinode)
 /*
  * vinode_rebuild_block_tree -- rebuilds runtime tree of blocks
  */
-static int
+int
 vinode_rebuild_block_tree(struct pmemfile_vinode *vinode)
 {
 	struct ctree *c = ctree_new();
@@ -161,7 +115,7 @@ vinode_rebuild_block_tree(struct pmemfile_vinode *vinode)
  * is_offset_in_block -- check if the given offset is in the range
  * specified by the block metadata
  */
-static bool
+bool
 is_offset_in_block(const struct pmemfile_block_desc *block, uint64_t offset)
 {
 	if (block == NULL)
@@ -189,7 +143,7 @@ is_block_data_initialized(const struct pmemfile_block_desc *block)
  * find_closest_block -- look up block metadata with the highest offset
  * lower than or equal to the offset argument
  */
-static struct pmemfile_block_desc *
+struct pmemfile_block_desc *
 find_closest_block(struct pmemfile_vinode *vinode, uint64_t off)
 {
 	return (void *)(uintptr_t)ctree_find_le(vinode->blocks, &off);
@@ -201,7 +155,7 @@ find_closest_block(struct pmemfile_vinode *vinode, uint64_t off)
  *
  * using the proposed last_block
  */
-static struct pmemfile_block_desc *
+struct pmemfile_block_desc *
 find_closest_block_with_hint(struct pmemfile_vinode *vinode, uint64_t offset,
 		struct pmemfile_block_desc *last_block)
 {
@@ -491,7 +445,7 @@ overallocate_size(uint64_t size)
  *
  * The remaining interval has zero size, ending the loop.
  */
-static void
+void
 vinode_allocate_interval(PMEMfilepool *pfp, struct pmemfile_vinode *vinode,
 		uint64_t offset, uint64_t size)
 {
@@ -668,8 +622,6 @@ find_following_block(struct pmemfile_vinode *vinode,
 		return vinode->first_block;
 }
 
-enum cpy_direction { read_from_blocks, write_to_blocks };
-
 /*
  * read_block_range - copy data to user supplied buffer
  */
@@ -746,7 +698,7 @@ write_block_range(PMEMfilepool *pfp, struct pmemfile_block_desc *block,
  * blocks to be already allocated. In case of reading, it is ok to skip holes
  * between blocks.
  */
-static struct pmemfile_block_desc *
+struct pmemfile_block_desc *
 iterate_on_file_range(PMEMfilepool *pfp, struct pmemfile_vinode *vinode,
 		struct pmemfile_block_desc *starting_block, uint64_t offset,
 		uint64_t len, char *buf, enum cpy_direction dir)
@@ -858,830 +810,6 @@ iterate_on_file_range(PMEMfilepool *pfp, struct pmemfile_vinode *vinode,
 	return last_block;
 }
 
-/*
- * vinode_write -- writes to file
- */
-static void
-vinode_write(PMEMfilepool *pfp, struct pmemfile_vinode *vinode, size_t offset,
-		struct pmemfile_block_desc **last_block,
-		const char *buf, size_t count)
-{
-	ASSERT(count > 0);
-	struct pmemfile_inode *inode = vinode->inode;
-
-	/*
-	 * Two steps:
-	 * - Zero Fill some new blocks, in case the file is extended by
-	 *   writing to the file after seeking past file size ( optionally )
-	 * - Copy the data from the users buffer
-	 */
-
-	uint64_t original_size = inode->size;
-	uint64_t new_size = inode->size;
-
-	if (offset + count > original_size)
-		new_size = offset + count;
-
-	/* All blocks needed for writing are properly allocated at this point */
-
-	struct pmemfile_block_desc *block =
-		find_closest_block_with_hint(vinode, offset, *last_block);
-
-	block = iterate_on_file_range(pfp, vinode, block, offset,
-			count, (char *)buf, write_to_blocks);
-
-	if (block)
-		*last_block = block;
-
-	if (new_size != original_size) {
-		TX_ADD_FIELD_DIRECT(inode, size);
-		inode->size = new_size;
-	}
-}
-
-static pmemfile_ssize_t
-pmemfile_pwritev_internal(PMEMfilepool *pfp,
-		struct pmemfile_vinode *vinode,
-		struct pmemfile_block_desc **last_block,
-		uint64_t file_flags,
-		size_t offset,
-		const pmemfile_iovec_t *iov,
-		int iovcnt)
-{
-	LOG(LDBG, "vinode %p iov %p iovcnt %d", vinode, iov, iovcnt);
-
-	if (!vinode_is_regular_file(vinode)) {
-		errno = EINVAL;
-		return -1;
-	}
-
-	if (!(file_flags & PFILE_WRITE)) {
-		errno = EBADF;
-		return -1;
-	}
-
-	if (iovcnt == 0)
-		return 0;
-
-	int error = 0;
-
-	struct pmemfile_inode *inode = vinode->inode;
-
-	os_rwlock_wrlock(&vinode->rwlock);
-
-	vinode_snapshot(vinode);
-
-	size_t ret = 0;
-
-	TX_BEGIN_CB(pfp->pop, cb_queue, pfp) {
-		if (!vinode->blocks) {
-			int err = vinode_rebuild_block_tree(vinode);
-			if (err)
-				pmemfile_tx_abort(err);
-		}
-
-		if (file_flags & PFILE_APPEND)
-			offset = inode->size;
-
-		size_t sum_len = 0;
-		for (int i = 0; i < iovcnt; ++i) {
-			size_t len = iov[i].iov_len;
-
-			if ((pmemfile_ssize_t)len < 0)
-				len = SSIZE_MAX;
-
-			if ((pmemfile_ssize_t)(sum_len + len) < 0)
-				len = SSIZE_MAX - ret;
-
-			/* overflow check */
-			if (offset + sum_len + len < offset)
-				len = SIZE_MAX - offset - sum_len;
-
-			sum_len += len;
-
-			if (len != iov[i].iov_len)
-				break;
-		}
-
-		if (sum_len > 0)
-			vinode_allocate_interval(pfp, vinode, offset, sum_len);
-
-		for (int i = 0; i < iovcnt; ++i) {
-			size_t len = iov[i].iov_len;
-
-			if ((pmemfile_ssize_t)len < 0)
-				len = SSIZE_MAX;
-
-			if ((pmemfile_ssize_t)(ret + len) < 0)
-				len = SSIZE_MAX - ret;
-
-			if (offset + len < offset) /* overflow check */
-				len = SIZE_MAX - offset;
-
-			if (len > 0)
-				vinode_write(pfp, vinode, offset, last_block,
-						iov[i].iov_base, len);
-
-			ret += len;
-			offset += len;
-
-			if (len != iov[i].iov_len)
-				break;
-		}
-
-		if (ret > 0) {
-			struct pmemfile_time tm;
-			get_current_time(&tm);
-			TX_SET_DIRECT(inode, mtime, tm);
-		}
-	} TX_ONABORT {
-		error = errno;
-		vinode_restore_on_abort(vinode);
-	} TX_END
-
-	os_rwlock_unlock(&vinode->rwlock);
-
-	if (error) {
-		errno = error;
-		return -1;
-	}
-
-	return (pmemfile_ssize_t)ret;
-}
-
-/*
- * pmemfile_write -- writes to file
- */
-pmemfile_ssize_t
-pmemfile_write(PMEMfilepool *pfp, PMEMfile *file, const void *buf, size_t count)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	os_mutex_lock(&file->mutex);
-
-	struct pmemfile_block_desc *last_block = file->block_pointer_cache;
-	pmemfile_iovec_t vec;
-	vec.iov_base = (void *)buf;
-	vec.iov_len = count;
-
-	pmemfile_ssize_t ret = pmemfile_pwritev_internal(pfp, file->vinode,
-			&last_block, file->flags, file->offset, &vec, 1);
-	if (ret >= 0) {
-		file->offset += (size_t)ret;
-		file->block_pointer_cache = last_block;
-	}
-
-	os_mutex_unlock(&file->mutex);
-
-	return ret;
-}
-
-pmemfile_ssize_t
-pmemfile_writev(PMEMfilepool *pfp, PMEMfile *file, const pmemfile_iovec_t *iov,
-		int iovcnt)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	os_mutex_lock(&file->mutex);
-
-	struct pmemfile_block_desc *last_block = file->block_pointer_cache;
-
-	pmemfile_ssize_t ret = pmemfile_pwritev_internal(pfp, file->vinode,
-			&last_block, file->flags, file->offset, iov, iovcnt);
-	if (ret >= 0) {
-		file->offset += (size_t)ret;
-		file->block_pointer_cache = last_block;
-	}
-
-	os_mutex_unlock(&file->mutex);
-
-	return ret;
-}
-
-pmemfile_ssize_t
-pmemfile_pwrite(PMEMfilepool *pfp, PMEMfile *file, const void *buf,
-		size_t count, pmemfile_off_t offset)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (offset < 0) {
-		errno = EINVAL;
-		return -1;
-	}
-
-	os_mutex_lock(&file->mutex);
-
-	struct pmemfile_block_desc *last_block = file->block_pointer_cache;
-	struct pmemfile_vinode *vinode = file->vinode;
-	uint64_t flags = file->flags;
-
-	os_mutex_unlock(&file->mutex);
-
-	pmemfile_iovec_t vec;
-	vec.iov_base = (void *)buf;
-	vec.iov_len = count;
-
-	return pmemfile_pwritev_internal(pfp, vinode, &last_block, flags,
-			(size_t)offset, &vec, 1);
-}
-
-pmemfile_ssize_t
-pmemfile_pwritev(PMEMfilepool *pfp, PMEMfile *file, const pmemfile_iovec_t *iov,
-		int iovcnt, pmemfile_off_t offset)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (offset < 0) {
-		errno = EINVAL;
-		return -1;
-	}
-
-	os_mutex_lock(&file->mutex);
-
-	struct pmemfile_block_desc *last_block = file->block_pointer_cache;
-	struct pmemfile_vinode *vinode = file->vinode;
-	uint64_t flags = file->flags;
-
-	os_mutex_unlock(&file->mutex);
-
-	return pmemfile_pwritev_internal(pfp, vinode, &last_block, flags,
-			(size_t)offset, iov, iovcnt);
-}
-
-/*
- * vinode_read -- reads file
- */
-static size_t
-vinode_read(PMEMfilepool *pfp, struct pmemfile_vinode *vinode, size_t offset,
-		struct pmemfile_block_desc **last_block, char *buf,
-		size_t count)
-{
-	uint64_t size = vinode->inode->size;
-
-	/*
-	 * Start reading at offset, stop reading
-	 * when end of file is reached, or count bytes were read.
-	 * The following two branches compute how many bytes are
-	 * going to be read.
-	 */
-	if (offset >= size)
-		return 0; /* EOF already */
-
-	if (size - offset < count)
-		count = size - offset;
-
-	struct pmemfile_block_desc *block =
-		find_closest_block_with_hint(vinode, offset, *last_block);
-
-	block = iterate_on_file_range(pfp, vinode, block, offset,
-			count, buf, read_from_blocks);
-
-	if (block)
-		*last_block = block;
-
-	return count;
-}
-
-static int
-time_cmp(const struct pmemfile_time *t1, const struct pmemfile_time *t2)
-{
-	if (t1->sec < t2->sec)
-		return -1;
-	if (t1->sec > t2->sec)
-		return 1;
-	if (t1->nsec < t2->nsec)
-		return -1;
-	if (t1->nsec > t2->nsec)
-		return 1;
-	return 0;
-}
-
-static pmemfile_ssize_t
-pmemfile_preadv_internal(PMEMfilepool *pfp,
-		struct pmemfile_vinode *vinode,
-		struct pmemfile_block_desc **last_block,
-		uint64_t file_flags,
-		size_t offset,
-		const pmemfile_iovec_t *iov,
-		int iovcnt)
-{
-	LOG(LDBG, "vinode %p iov %p iovcnt %d", vinode, iov, iovcnt);
-
-	if (!vinode_is_regular_file(vinode)) {
-		errno = EINVAL;
-		return -1;
-	}
-
-	if (!(file_flags & PFILE_READ)) {
-		errno = EBADF;
-		return -1;
-	}
-
-	if (iovcnt == 0)
-		return 0;
-
-	struct pmemfile_inode *inode = vinode->inode;
-
-	/*
-	 * We want read to be performed under read lock, but we need the block
-	 * tree to exist. If it doesn't exist we have to drop the lock we hold,
-	 * take it in write mode (because other thread may want to do the same),
-	 * check that it doesn't exist (another thread may already did that),
-	 * drop the lock again, take it in read mode and check AGAIN (because
-	 * another thread may have destroyed the block tree while we weren't
-	 * holding the lock).
-	 */
-	os_rwlock_rdlock(&vinode->rwlock);
-	while (!vinode->blocks) {
-		os_rwlock_unlock(&vinode->rwlock);
-		os_rwlock_wrlock(&vinode->rwlock);
-
-		int err = 0;
-		if (!vinode->blocks)
-			err = vinode_rebuild_block_tree(vinode);
-		os_rwlock_unlock(&vinode->rwlock);
-
-		if (err) {
-			errno = err;
-			return -1;
-		}
-
-		os_rwlock_rdlock(&vinode->rwlock);
-	}
-
-	size_t ret = 0;
-
-	for (int i = 0; i < iovcnt; ++i) {
-		size_t len = iov[i].iov_len;
-		if ((pmemfile_ssize_t)(ret + len) < 0)
-			len = SSIZE_MAX - ret;
-		ASSERT((pmemfile_ssize_t)(ret + len) >= 0);
-
-		size_t bytes_read = vinode_read(pfp, vinode, offset, last_block,
-				iov[i].iov_base, len);
-
-		ret += bytes_read;
-		offset += bytes_read;
-		if (bytes_read != len)
-			break;
-	}
-
-	bool update_atime = !(file_flags & PFILE_NOATIME);
-	struct pmemfile_time tm;
-
-	if (update_atime) {
-		struct pmemfile_time tm1d;
-		get_current_time(&tm);
-		tm1d.nsec = tm.nsec;
-		tm1d.sec = tm.sec - 86400;
-
-		/* relatime */
-		update_atime =	time_cmp(&inode->atime, &tm1d) < 0 ||
-				time_cmp(&inode->atime, &inode->ctime) < 0 ||
-				time_cmp(&inode->atime, &inode->mtime) < 0;
-	}
-
-	os_rwlock_unlock(&vinode->rwlock);
-
-	if (update_atime) {
-		os_rwlock_wrlock(&vinode->rwlock);
-
-		TX_BEGIN_CB(pfp->pop, cb_queue, pfp) {
-			TX_SET_DIRECT(inode, atime, tm);
-		} TX_ONABORT {
-			LOG(LINF, "can not update inode atime");
-		} TX_END
-
-		os_rwlock_unlock(&vinode->rwlock);
-	}
-
-	return (pmemfile_ssize_t)ret;
-}
-
-/*
- * pmemfile_read -- reads file
- */
-pmemfile_ssize_t
-pmemfile_read(PMEMfilepool *pfp, PMEMfile *file, void *buf, size_t count)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	os_mutex_lock(&file->mutex);
-
-	struct pmemfile_block_desc *last_block = file->block_pointer_cache;
-	pmemfile_iovec_t vec;
-	vec.iov_base = buf;
-	vec.iov_len = count;
-
-	pmemfile_ssize_t ret = pmemfile_preadv_internal(pfp, file->vinode,
-			&last_block, file->flags, file->offset, &vec, 1);
-	if (ret >= 0) {
-		file->offset += (size_t)ret;
-		file->block_pointer_cache = last_block;
-	}
-
-	os_mutex_unlock(&file->mutex);
-
-	return ret;
-}
-
-pmemfile_ssize_t
-pmemfile_readv(PMEMfilepool *pfp, PMEMfile *file, const pmemfile_iovec_t *iov,
-		int iovcnt)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	os_mutex_lock(&file->mutex);
-
-	struct pmemfile_block_desc *last_block = file->block_pointer_cache;
-
-	pmemfile_ssize_t ret = pmemfile_preadv_internal(pfp, file->vinode,
-			&last_block, file->flags, file->offset, iov, iovcnt);
-	if (ret >= 0) {
-		file->offset += (size_t)ret;
-		file->block_pointer_cache = last_block;
-	}
-
-	os_mutex_unlock(&file->mutex);
-
-	return ret;
-}
-
-pmemfile_ssize_t
-pmemfile_pread(PMEMfilepool *pfp, PMEMfile *file, void *buf, size_t count,
-		pmemfile_off_t offset)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (offset < 0) {
-		errno = EINVAL;
-		return -1;
-	}
-
-	os_mutex_lock(&file->mutex);
-
-	struct pmemfile_block_desc *last_block = file->block_pointer_cache;
-	struct pmemfile_vinode *vinode = file->vinode;
-	uint64_t flags = file->flags;
-
-	os_mutex_unlock(&file->mutex);
-
-	pmemfile_iovec_t vec;
-	vec.iov_base = buf;
-	vec.iov_len = count;
-
-	return pmemfile_preadv_internal(pfp, vinode, &last_block, flags,
-			(size_t)offset, &vec, 1);
-}
-
-
-pmemfile_ssize_t
-pmemfile_preadv(PMEMfilepool *pfp, PMEMfile *file, const pmemfile_iovec_t *iov,
-		int iovcnt, pmemfile_off_t offset)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (offset < 0) {
-		errno = EINVAL;
-		return -1;
-	}
-
-	os_mutex_lock(&file->mutex);
-
-	struct pmemfile_block_desc *last_block = file->block_pointer_cache;
-	struct pmemfile_vinode *vinode = file->vinode;
-	uint64_t flags = file->flags;
-
-	os_mutex_unlock(&file->mutex);
-
-	return pmemfile_preadv_internal(pfp, vinode, &last_block, flags,
-			(size_t)offset, iov, iovcnt);
-}
-
-/*
- * lseek_seek_data -- part of the lseek implementation
- * Looks for data (not a hole), starting at the specified offset.
- */
-static pmemfile_off_t
-lseek_seek_data(struct pmemfile_vinode *vinode, pmemfile_off_t offset,
-		pmemfile_off_t fsize)
-{
-	if (vinode->blocks == NULL) {
-		int err = vinode_rebuild_block_tree(vinode);
-		if (err)
-			return err;
-	}
-
-	struct pmemfile_block_desc *block =
-			find_closest_block(vinode, (uint64_t)offset);
-	if (block == NULL) {
-		/* offset is before the first block */
-		if (vinode->first_block == NULL)
-			return fsize; /* No data in the whole file */
-		else
-			return (pmemfile_off_t)vinode->first_block->offset;
-	}
-
-	if (is_offset_in_block(block, (uint64_t)offset))
-		return offset;
-
-	block = D_RW(block->next);
-
-	if (block == NULL)
-		return fsize; /* No more data in file */
-
-	return (pmemfile_off_t)block->offset;
-}
-
-/*
- * lseek_seek_hole -- part of the lseek implementation
- * Looks for a hole, starting at the specified offset.
- */
-static pmemfile_off_t
-lseek_seek_hole(struct pmemfile_vinode *vinode, pmemfile_off_t offset,
-		pmemfile_off_t fsize)
-{
-	if (vinode->blocks == NULL) {
-		int err = vinode_rebuild_block_tree(vinode);
-		if (err)
-			return err;
-	}
-
-	struct pmemfile_block_desc *block =
-			find_closest_block(vinode, (uint64_t)offset);
-
-	while (block != NULL && offset < fsize) {
-		pmemfile_off_t block_end =
-				(pmemfile_off_t)block->offset + block->size;
-
-		struct pmemfile_block_desc *next = D_RW(block->next);
-
-		if (block_end >= offset)
-			offset = block_end; /* seek to the end of block */
-
-		if (next == NULL)
-			break; /* the rest of the file is treated as a hole */
-		else if (offset < (pmemfile_off_t)next->offset)
-			break; /* offset is in a hole between two blocks */
-
-		block = next;
-	}
-
-	return offset;
-}
-
-/*
- * lseek_seek_data_or_hole -- part of the lseek implementation
- * Expects the vinode to be locked while being called.
- */
-static pmemfile_off_t
-lseek_seek_data_or_hole(struct pmemfile_vinode *vinode, pmemfile_off_t offset,
-			int whence)
-{
-	pmemfile_ssize_t fsize = (pmemfile_ssize_t)vinode->inode->size;
-
-	if (!vinode_is_regular_file(vinode))
-		return -EBADF; /* XXX directories are not supported here yet */
-
-	if (offset > fsize) {
-		/*
-		 * From GNU man page: ENXIO if
-		 * "...ENXIO  whence is SEEK_DATA or SEEK_HOLE, and the file
-		 * offset is beyond the end of the file..."
-		 */
-		return -ENXIO;
-	}
-
-	if (offset < 0) /* this seems to be allowed by POSIX and Linux */
-		offset = 0;
-
-	if (whence == PMEMFILE_SEEK_DATA) {
-		offset = lseek_seek_data(vinode, offset, fsize);
-	} else {
-		ASSERT(whence == PMEMFILE_SEEK_HOLE);
-		offset = lseek_seek_hole(vinode, offset, fsize);
-	}
-
-	if (offset > fsize)
-		offset = fsize;
-
-	return offset;
-}
-
-/*
- * pmemfile_lseek_locked -- changes file current offset
- */
-static pmemfile_off_t
-pmemfile_lseek_locked(PMEMfilepool *pfp, PMEMfile *file, pmemfile_off_t offset,
-		int whence)
-{
-	(void) pfp;
-
-	LOG(LDBG, "file %p offset %ld whence %d", file, offset, whence);
-
-	if (file->flags & PFILE_PATH) {
-		errno = EBADF;
-		return -1;
-	}
-
-	if (vinode_is_dir(file->vinode)) {
-		if (whence == PMEMFILE_SEEK_END) {
-			errno = EINVAL;
-			return -1;
-		}
-	} else if (vinode_is_regular_file(file->vinode)) {
-		/* Nothing to do for now */
-	} else {
-		errno = EINVAL;
-		return -1;
-	}
-
-	struct pmemfile_vinode *vinode = file->vinode;
-	struct pmemfile_inode *inode = vinode->inode;
-	pmemfile_off_t ret;
-	int new_errno = EINVAL;
-
-	switch (whence) {
-		case PMEMFILE_SEEK_SET:
-			ret = offset;
-			if (ret < 0) {
-				/*
-				 * From POSIX: EINVAL if
-				 * "...the resulting file offset would be
-				 * negative for a regular file..."
-				 */
-				new_errno = EINVAL;
-			}
-			break;
-		case PMEMFILE_SEEK_CUR:
-			ret = (pmemfile_off_t)file->offset + offset;
-			if (ret < 0) {
-				if (offset < 0) {
-					new_errno = EINVAL;
-				} else {
-					/*
-					 * From POSIX: EOVERFLOW if
-					 * "...The resulting file offset would
-					 * be a value which cannot be
-					 * represented correctly in an object
-					 * of type off_t..."
-					 */
-					new_errno = EOVERFLOW;
-				}
-			}
-			break;
-		case PMEMFILE_SEEK_END:
-			os_rwlock_rdlock(&vinode->rwlock);
-			ret = (pmemfile_off_t)inode->size + offset;
-			os_rwlock_unlock(&vinode->rwlock);
-			if (ret < 0) {
-				/* Errors as in SEEK_CUR */
-				if (offset < 0)
-					new_errno = EINVAL;
-				else
-					new_errno = EOVERFLOW;
-			}
-			break;
-		case PMEMFILE_SEEK_DATA:
-		case PMEMFILE_SEEK_HOLE:
-			/*
-			 * We may need to rebuild the block tree, so we have to
-			 * take vinode lock in write mode.
-			 */
-			os_rwlock_wrlock(&vinode->rwlock);
-			ret = lseek_seek_data_or_hole(vinode, offset, whence);
-			if (ret < 0) {
-				new_errno = (int)-ret;
-				ret = -1;
-			}
-			os_rwlock_unlock(&vinode->rwlock);
-			break;
-		default:
-			ret = -1;
-			break;
-	}
-
-	if (ret < 0) {
-		ret = -1;
-		errno = new_errno;
-	} else {
-		if (file->offset != (size_t)ret)
-			LOG(LDBG, "off diff: old %lu != new %lu", file->offset,
-					(size_t)ret);
-		file->offset = (size_t)ret;
-	}
-
-	return ret;
-}
-
-/*
- * pmemfile_lseek -- changes file current offset
- */
-pmemfile_off_t
-pmemfile_lseek(PMEMfilepool *pfp, PMEMfile *file, pmemfile_off_t offset,
-		int whence)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	COMPILE_ERROR_ON(sizeof(offset) != 8);
-	pmemfile_off_t ret;
-
-	os_mutex_lock(&file->mutex);
-	ret = pmemfile_lseek_locked(pfp, file, offset, whence);
-	os_mutex_unlock(&file->mutex);
-
-	return ret;
-}
 
 /*
  * is_block_contained_by_interval -- see vinode_remove_interval
@@ -1761,7 +889,7 @@ is_block_at_right_edge(struct pmemfile_block_desc *block,
  * At this point, the file size is not changed, but the corresponding file
  * contents would remain zero bytes, if they were not snapshotted.
  */
-static void
+void
 vinode_remove_interval(struct pmemfile_vinode *vinode,
 			uint64_t offset, uint64_t len)
 {
@@ -1847,130 +975,5 @@ vinode_remove_interval(struct pmemfile_vinode *vinode,
 
 			block = D_RW(block->prev);
 		}
-	}
-}
-
-/*
- * vinode_truncate -- changes file size to size
- *
- * Should only be called inside pmemobj transactions.
- */
-void
-vinode_truncate(PMEMfilepool *pfp, struct pmemfile_vinode *vinode,
-		uint64_t size)
-{
-	struct pmemfile_inode *inode = vinode->inode;
-
-	ASSERTeq(pmemobj_tx_stage(), TX_STAGE_WORK);
-
-	if (vinode->blocks == NULL) {
-		int err = vinode_rebuild_block_tree(vinode);
-		if (err)
-			pmemfile_tx_abort(err);
-	}
-
-	cb_push_front(TX_STAGE_ONABORT,
-		(cb_basic)vinode_destroy_data_state,
-		vinode);
-
-	/*
-	 * Might need to handle the special case where size == 0.
-	 * Setting all the next and prev fields is pointless, when all the
-	 * blocks are removed.
-	 */
-	vinode_remove_interval(vinode, size, UINT64_MAX - size);
-	if (inode->size < size)
-		vinode_allocate_interval(pfp, vinode,
-		    inode->size, size - inode->size);
-
-	if (inode->size != size) {
-		TX_ADD_DIRECT(&inode->size);
-		inode->size = size;
-
-		struct pmemfile_time tm;
-		get_current_time(&tm);
-		TX_SET_DIRECT(inode, mtime, tm);
-	}
-}
-
-int
-vinode_fallocate(PMEMfilepool *pfp, struct pmemfile_vinode *vinode, int mode,
-		uint64_t offset, uint64_t length)
-{
-	int error = 0;
-
-	if (!vinode_is_regular_file(vinode))
-		return EBADF;
-
-	uint64_t off_plus_len = offset + length;
-
-	if (mode & PMEMFILE_FALLOC_FL_PUNCH_HOLE)
-		narrow_to_full_pages(&offset, &length);
-	else
-		expand_to_full_pages(&offset, &length);
-
-	if (length == 0)
-		return 0;
-
-	vinode_snapshot(vinode);
-
-	if (vinode->blocks == NULL) {
-		error = vinode_rebuild_block_tree(vinode);
-		if (error)
-			return error;
-	}
-
-	TX_BEGIN_CB(pfp->pop, cb_queue, pfp) {
-		if (mode & PMEMFILE_FALLOC_FL_PUNCH_HOLE) {
-			ASSERT(mode & PMEMFILE_FALLOC_FL_KEEP_SIZE);
-			vinode_remove_interval(vinode, offset, length);
-		} else {
-			vinode_allocate_interval(pfp, vinode, offset, length);
-			if ((mode & PMEMFILE_FALLOC_FL_KEEP_SIZE) == 0) {
-				if (vinode->inode->size < off_plus_len) {
-					TX_ADD_DIRECT(&vinode->inode->size);
-					vinode->inode->size = off_plus_len;
-				}
-			}
-		}
-	} TX_ONABORT {
-		error = errno;
-		vinode_restore_on_abort(vinode);
-	} TX_END
-
-	return error;
-}
-
-/*
- * vinode_snapshot
- * Saves some volatile state in vinode, that can be altered during a
- * transaction. These volatile data are not restored by pmemobj upon
- * transaction abort.
- */
-void
-vinode_snapshot(struct pmemfile_vinode *vinode)
-{
-	vinode->snapshot.first_free_block = vinode->first_free_block;
-	vinode->snapshot.first_block = vinode->first_block;
-}
-
-/*
- * vinode_restore_on_abort - vinode_snapshot's counterpart
- * This must be added to abort handlers, where vinode_snapshot was
- * called in the transaction.
- */
-void
-vinode_restore_on_abort(struct pmemfile_vinode *vinode)
-{
-	vinode->first_free_block = vinode->snapshot.first_free_block;
-	vinode->first_block = vinode->snapshot.first_block;
-
-	/*
-	 * The ctree is not restored here. It is rebuilt the next
-	 * time the vinode is used.
-	 */
-	if (vinode->blocks) {
-		ctree_delete(vinode->blocks);
-		vinode->blocks = NULL;
 	}
 }
