+++ src/libpmemfile-posix/data.h
@@ -32,7 +32,7 @@
 #ifndef PMEMFILE_DATA_H
 #define PMEMFILE_DATA_H
 
-#include "file.h"
+#include "inode.h"
 
 extern size_t pmemfile_posix_block_size;
 extern bool pmemfile_overallocate_on_append;
@@ -41,7 +41,27 @@ void vinode_destroy_data_state(PMEMfilepool *pfp,
 			struct pmemfile_vinode *vinode);
 void vinode_truncate(PMEMfilepool *pfp, struct pmemfile_vinode *vinode,
 			uint64_t size);
-int vinode_fallocate(PMEMfilepool *pfp, struct pmemfile_vinode *vinode,
-			int mode, uint64_t offset, uint64_t length);
+
+int vinode_rebuild_block_tree(struct pmemfile_vinode *vinode);
+void vinode_remove_interval(struct pmemfile_vinode *vinode,
+			uint64_t offset, uint64_t len);
+void vinode_allocate_interval(PMEMfilepool *pfp, struct pmemfile_vinode *vinode,
+		uint64_t offset, uint64_t size);
+
+struct pmemfile_block_desc *find_closest_block(struct pmemfile_vinode *vinode,
+		uint64_t off);
+struct pmemfile_block_desc *find_closest_block_with_hint(
+		struct pmemfile_vinode *vinode, uint64_t offset,
+		struct pmemfile_block_desc *last_block);
+
+bool is_offset_in_block(const struct pmemfile_block_desc *block,
+		uint64_t offset);
+
+enum cpy_direction { read_from_blocks, write_to_blocks };
+
+struct pmemfile_block_desc *iterate_on_file_range(PMEMfilepool *pfp,
+		struct pmemfile_vinode *vinode,
+		struct pmemfile_block_desc *starting_block, uint64_t offset,
+		uint64_t len, char *buf, enum cpy_direction dir);
 
 #endif
