+++ src/libpmemfile-posix/dir.c
@@ -234,94 +234,6 @@ inode_add_dirent(PMEMfilepool *pfp,
 }
 
 /*
- * vinode_update_parent -- update .. entry of "vinode" from "src_parent" to
- * "dst_parent"
- */
-void
-vinode_update_parent(PMEMfilepool *pfp,
-		struct pmemfile_vinode *vinode,
-		struct pmemfile_vinode *src_parent,
-		struct pmemfile_vinode *dst_parent)
-{
-	ASSERTeq(pmemobj_tx_stage(), TX_STAGE_WORK);
-
-	struct pmemfile_dir *dir = &vinode->inode->file_data.dir;
-
-	struct pmemfile_dirent *dirent = NULL;
-
-	do {
-		for (uint32_t i = 0; i < dir->num_elements; ++i) {
-			if (strcmp(dir->dirents[i].name, "..") == 0) {
-				dirent = &dir->dirents[i];
-				break;
-			}
-		}
-
-		if (dirent)
-			break;
-
-		dir = D_RW(dir->next);
-	} while (dir);
-
-	ASSERTne(dirent, NULL);
-	ASSERT(TOID_EQUALS(dirent->inode, src_parent->tinode));
-	ASSERTeq(vinode->parent, src_parent);
-
-	TX_ADD_DIRECT(&src_parent->inode->nlink);
-	src_parent->inode->nlink--;
-
-	TX_ADD_DIRECT(&dst_parent->inode->nlink);
-	dst_parent->inode->nlink++;
-
-	TX_ADD_DIRECT(&dirent->inode);
-	dirent->inode = dst_parent->tinode;
-
-	vinode->parent = vinode_ref(pfp, dst_parent);
-}
-
-/*
- * vinode_new_dir -- creates new directory relative to parent
- *
- * Note: caller must hold WRITE lock on parent.
- * Must be called in a transaction.
- */
-TOID(struct pmemfile_inode)
-vinode_new_dir(PMEMfilepool *pfp, struct pmemfile_vinode *parent,
-		const char *name, size_t namelen, struct pmemfile_cred *cred,
-		pmemfile_mode_t mode)
-{
-	LOG(LDBG, "parent 0x%" PRIx64 " ppath %s new_name %.*s",
-			parent ? parent->tinode.oid.off : 0,
-			pmfi_path(parent), (int)namelen, name);
-
-	ASSERTeq(pmemobj_tx_stage(), TX_STAGE_WORK);
-
-	if (mode & ~(pmemfile_mode_t)PMEMFILE_ACCESSPERMS) {
-		/* XXX: what the kernel does in this case? */
-		ERR("invalid mode flags 0%o", mode);
-		pmemfile_tx_abort(EINVAL);
-	}
-
-	TOID(struct pmemfile_inode) tchild =
-			inode_alloc(pfp, cred, PMEMFILE_S_IFDIR | mode);
-	struct pmemfile_inode *child = D_RW(tchild);
-	struct pmemfile_time t = child->ctime;
-
-	/* add . and .. to new directory */
-	inode_add_dirent(pfp, tchild, ".", 1, tchild, t);
-
-	if (parent == NULL) { /* special case - root directory */
-		inode_add_dirent(pfp, tchild, "..", 2, tchild, t);
-	} else {
-		inode_add_dirent(pfp, tchild, "..", 2, parent->tinode, t);
-		inode_add_dirent(pfp, parent->tinode, name, namelen, tchild,
-				t);
-	}
-
-	return tchild;
-}
-
-/*
  * vinode_lookup_dirent_by_name_locked -- looks up file name in passed directory
  *
  * Caller must hold lock on parent.
@@ -466,371 +378,6 @@ end:
 	os_rwlock_unlock(&parent->rwlock);
 
 	return vinode;
-}
-
-/*
- * vinode_unlink_file -- removes file dirent from directory
- *
- * Must be called in a transaction. Caller must have exclusive access to both
- * parent and child inode by locking them in WRITE mode.
- */
-void
-vinode_unlink_file(PMEMfilepool *pfp,
-		struct pmemfile_vinode *parent,
-		struct pmemfile_dirent *dirent,
-		struct pmemfile_vinode *vinode)
-{
-	LOG(LDBG, "parent 0x%" PRIx64 " ppath %s name %s",
-		parent->tinode.oid.off, pmfi_path(parent), dirent->name);
-
-	ASSERTeq(pmemobj_tx_stage(), TX_STAGE_WORK);
-
-	TOID(struct pmemfile_inode) tinode = dirent->inode;
-	struct pmemfile_inode *inode = D_RW(tinode);
-
-	ASSERT(inode->nlink > 0);
-
-	TX_ADD_FIELD(tinode, nlink);
-	/*
-	 * Snapshot inode and the first byte of a name (because we are going
-	 * to overwrite just one byte) using one call.
-	 */
-	pmemobj_tx_add_range_direct(dirent, sizeof(dirent->inode) + 1);
-
-	struct pmemfile_time tm;
-	get_current_time(&tm);
-
-	if (--inode->nlink > 0) {
-		/*
-		 * From "stat" man page:
-		 * "The field st_ctime is changed by writing or by setting inode
-		 * information (i.e., owner, group, link count, mode, etc.)."
-		 */
-		TX_SET_DIRECT(vinode->inode, ctime, tm);
-	}
-	/*
-	 * From "stat" man page:
-	 * "st_mtime of a directory is changed by the creation
-	 * or deletion of files in that directory."
-	 */
-	TX_SET_DIRECT(parent->inode, mtime, tm);
-
-	dirent->name[0] = '\0';
-	dirent->inode = TOID_NULL(struct pmemfile_inode);
-}
-
-#define DIRENT_ID_MASK 0xffffffffULL
-
-#define DIR_ID(offset) ((offset) >> 32)
-#define DIRENT_ID(offset) ((offset) & DIRENT_ID_MASK)
-
-/*
- * file_seek_dir - translates between file->offset and dir/dirent
- *
- * returns 0 on EOF
- * returns !0 on successful translation
- */
-static int
-file_seek_dir(PMEMfile *file, struct pmemfile_dir **dir, unsigned *dirent)
-{
-	struct pmemfile_inode *inode = file->vinode->inode;
-
-	if (file->offset == 0) {
-		file->dir_pos.dir = &inode->file_data.dir;
-		file->dir_pos.dir_id = 0;
-
-		*dir = file->dir_pos.dir;
-	} else if (DIR_ID(file->offset) == file->dir_pos.dir_id) {
-		*dir = file->dir_pos.dir;
-		if (*dir == NULL)
-			return 0;
-	} else {
-		*dir = &inode->file_data.dir;
-
-		unsigned dir_id = 0;
-		while (DIR_ID(file->offset) != dir_id) {
-			if (TOID_IS_NULL((*dir)->next))
-				return 0;
-			*dir = D_RW((*dir)->next);
-			++dir_id;
-		}
-
-		file->dir_pos.dir = *dir;
-		file->dir_pos.dir_id = dir_id;
-	}
-	*dirent = DIRENT_ID(file->offset);
-
-	while (*dirent >= (*dir)->num_elements) {
-		if (TOID_IS_NULL((*dir)->next))
-			return 0;
-
-		*dirent -= (*dir)->num_elements;
-		*dir = D_RW((*dir)->next);
-
-		file->dir_pos.dir = *dir;
-		file->dir_pos.dir_id++;
-	}
-
-	file->offset = ((size_t)file->dir_pos.dir_id) << 32 | *dirent;
-
-	return 1;
-}
-
-/*
- * inode_type - returns inode type, as returned by getdents
- */
-static char
-inode_type(const struct pmemfile_inode *inode)
-{
-	if (inode_is_regular_file(inode))
-		return PMEMFILE_DT_REG;
-
-	if (inode_is_symlink(inode))
-		return PMEMFILE_DT_LNK;
-
-	if (inode_is_dir(inode))
-		return PMEMFILE_DT_DIR;
-
-	ASSERT(0);
-	return PMEMFILE_DT_UNKNOWN;
-}
-
-/*
- * align_dirent_size - aligns dirent size to 8 bytes and returns alignment
- */
-static unsigned short
-align_dirent_size(unsigned short *slen)
-{
-	unsigned short alignment = (unsigned short)(8 - (*slen & 7));
-	if (alignment == 8)
-		alignment = 0;
-	*slen = (unsigned short)(*slen + alignment);
-	return alignment;
-}
-
-/*
- * get_next_dirent_off - returns (lseek) offset of next directory entry
- */
-static uint64_t
-get_next_dirent_off(PMEMfile *file, struct pmemfile_dir *dir,
-		unsigned dirent_id)
-{
-	uint64_t next_off = file->offset + 1;
-	if (dirent_id + 1 >= dir->num_elements)
-		next_off = ((next_off >> 32) + 1) << 32;
-	return next_off;
-}
-
-/*
- * fill_dirent32 -- fills data with dirent information using 32-bit getdents ABI
- */
-static unsigned short
-fill_dirent32(struct pmemfile_dirent *dirent, uint64_t next_off, unsigned left,
-		char *data)
-{
-	size_t namelen = strlen(dirent->name);
-	/* minimum size required */
-	unsigned short slen = (unsigned short)
-			(8 /* sizeof d_ino */ +
-			8 /* sizeof d_off */ +
-			2 /* sizeof d_reclen */ +
-			namelen + 1 /* strlen(d_name) + 1 */ +
-			1 /* sizeof d_type */);
-	/* add for the whole structure to be 8 bytes aligned */
-	unsigned short alignment = align_dirent_size(&slen);
-
-	if (slen > left)
-		return 0;
-
-	COMPILE_ERROR_ON(sizeof(dirent->inode.oid.off) != 8);
-	memcpy(data, &dirent->inode.oid.off, 8);
-	data += 8;
-
-	COMPILE_ERROR_ON(sizeof(next_off) != 8);
-	memcpy(data, &next_off, 8);
-	data += 8;
-
-	COMPILE_ERROR_ON(sizeof(slen) != 2);
-	memcpy(data, &slen, 2);
-	data += 2;
-
-	memcpy(data, dirent->name, namelen + 1);
-	data += namelen + 1;
-
-	while (alignment--)
-		*data++ = 0;
-
-	COMPILE_ERROR_ON(sizeof(inode_type(D_RO(dirent->inode))) != 1);
-	*data++ = inode_type(D_RO(dirent->inode));
-
-	return slen;
-}
-
-/*
- * fill_dirent64 -- fills data with dirent information using 64-bit getdents ABI
- */
-static unsigned short
-fill_dirent64(struct pmemfile_dirent *dirent, uint64_t next_off, unsigned left,
-		char *data)
-{
-	size_t namelen = strlen(dirent->name);
-	/* minimum size required */
-	unsigned short slen = (unsigned short)
-			(8 /* sizeof d_ino */ +
-			8 /* sizeof d_off */ +
-			2 /* sizeof d_reclen */ +
-			1 /* sizeof d_type */ +
-			namelen + 1 /* strlen(d_name) + 1 */);
-	/* add for the whole structure to be 8 bytes aligned */
-	unsigned short alignment = align_dirent_size(&slen);
-
-	if (slen > left)
-		return 0;
-
-	COMPILE_ERROR_ON(sizeof(dirent->inode.oid.off) != 8);
-	memcpy(data, &dirent->inode.oid.off, 8);
-	data += 8;
-
-	COMPILE_ERROR_ON(sizeof(next_off) != 8);
-	memcpy(data, &next_off, 8);
-	data += 8;
-
-	COMPILE_ERROR_ON(sizeof(slen) != 2);
-	memcpy(data, &slen, 2);
-	data += 2;
-
-	COMPILE_ERROR_ON(sizeof(inode_type(D_RO(dirent->inode))) != 1);
-	*data++ = inode_type(D_RO(dirent->inode));
-
-	memcpy(data, dirent->name, namelen + 1);
-	data += namelen + 1;
-
-	while (alignment--)
-		*data++ = 0;
-
-	return slen;
-}
-
-typedef unsigned short (*fill_dirent_type)(struct pmemfile_dirent *dirent,
-		uint64_t next_off, unsigned left, char *data);
-
-/*
- * pmemfile_getdents_worker -- traverses directory and fills dirent information
- */
-static int
-pmemfile_getdents_worker(PMEMfile *file, char *data, unsigned count,
-		fill_dirent_type fill_dirent)
-{
-	struct pmemfile_dir *dir;
-	unsigned dirent_id;
-
-	if (file_seek_dir(file, &dir, &dirent_id) == 0)
-		return 0;
-
-	int read1 = 0;
-
-	while (true) {
-		if (dirent_id >= dir->num_elements) {
-			if (TOID_IS_NULL(dir->next))
-				break;
-
-			dir = D_RW(dir->next);
-			file->dir_pos.dir = dir;
-			file->dir_pos.dir_id++;
-			dirent_id = 0;
-			file->offset = ((size_t)file->dir_pos.dir_id) << 32 | 0;
-		}
-		ASSERT(dir != NULL);
-
-		struct pmemfile_dirent *dirent = &dir->dirents[dirent_id];
-		if (TOID_IS_NULL(dirent->inode)) {
-			++dirent_id;
-			++file->offset;
-			continue;
-		}
-
-		uint64_t next_off = get_next_dirent_off(file, dir, dirent_id);
-
-		unsigned short slen = fill_dirent(dirent, next_off,
-				count - (unsigned)read1, data);
-		if (slen == 0)
-			break;
-
-		data += slen;
-		read1 += slen;
-
-		++dirent_id;
-		++file->offset;
-	}
-
-	return read1;
-}
-
-/*
- * pmemfile_getdents_generic -- generic implementation of pmemfile_getdents
- * which allows caller to pick ABI (fill_dirent)
- */
-static int
-pmemfile_getdents_generic(PMEMfilepool *pfp, PMEMfile *file, char *data,
-		unsigned count, fill_dirent_type fill_dirent)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!file) {
-		LOG(LUSR, "NULL file");
-		errno = EFAULT;
-		return -1;
-	}
-
-	struct pmemfile_vinode *vinode = file->vinode;
-
-	ASSERT(vinode != NULL);
-	if (!vinode_is_dir(vinode)) {
-		errno = ENOTDIR;
-		return -1;
-	}
-
-	if (!(file->flags & PFILE_READ)) {
-		errno = EBADF;
-		return -1;
-	}
-
-	if ((int)count < 0)
-		count = INT_MAX;
-
-	int bytes_read = 0;
-
-	os_mutex_lock(&file->mutex);
-	os_rwlock_rdlock(&vinode->rwlock);
-
-	bytes_read = pmemfile_getdents_worker(file, data, count, fill_dirent);
-	ASSERT(bytes_read >= 0);
-
-	os_rwlock_unlock(&vinode->rwlock);
-	os_mutex_unlock(&file->mutex);
-
-	ASSERT((unsigned)bytes_read <= count);
-	return bytes_read;
-}
-
-int
-pmemfile_getdents(PMEMfilepool *pfp, PMEMfile *file,
-			struct linux_dirent *dirp, unsigned count)
-{
-	return pmemfile_getdents_generic(pfp, file, (char *)dirp, count,
-			fill_dirent32);
-}
-
-int
-pmemfile_getdents64(PMEMfilepool *pfp, PMEMfile *file,
-			struct linux_dirent64 *dirp, unsigned count)
-{
-	return pmemfile_getdents_generic(pfp, file, (char *)dirp, count,
-			fill_dirent64);
 }
 
 static void
@@ -1038,194 +585,6 @@ path_info_cleanup(PMEMfilepool *pfp, struct pmemfile_path_info *path_info)
 	memset(path_info, 0, sizeof(*path_info));
 }
 
-static int
-_pmemfile_mkdirat(PMEMfilepool *pfp, struct pmemfile_vinode *dir,
-		const char *path, pmemfile_mode_t mode)
-{
-	struct pmemfile_path_info info;
-	struct pmemfile_cred cred;
-	if (cred_acquire(pfp, &cred))
-		return -1;
-
-	resolve_pathat(pfp, &cred, dir, path, &info, 0);
-
-	struct pmemfile_vinode *parent = info.parent;
-	int error = 0;
-
-	if (info.error) {
-		error = info.error;
-		goto end;
-	}
-
-	size_t namelen = component_length(info.remaining);
-
-	/* mkdir("/") */
-	if (namelen == 0) {
-		ASSERT(parent == pfp->root);
-		error = EEXIST;
-		goto end;
-	}
-
-	os_rwlock_wrlock(&parent->rwlock);
-
-	TX_BEGIN_CB(pfp->pop, cb_queue, pfp) {
-		if (!_vinode_can_access(&cred, parent, PFILE_WANT_WRITE))
-			pmemfile_tx_abort(EACCES);
-
-		vinode_new_dir(pfp, parent, info.remaining, namelen, &cred,
-				mode);
-	} TX_ONABORT {
-		error = errno;
-	} TX_END
-
-	os_rwlock_unlock(&parent->rwlock);
-
-end:
-	path_info_cleanup(pfp, &info);
-	cred_release(&cred);
-
-	if (error) {
-		errno = error;
-		return -1;
-	}
-
-	return 0;
-}
-
-void
-vinode_cleanup(PMEMfilepool *pfp, struct pmemfile_vinode *vinode,
-		bool preserve_errno)
-{
-	int error;
-	if (preserve_errno)
-		error = errno;
-
-	vinode_unref(pfp, vinode);
-
-	if (preserve_errno)
-		errno = error;
-}
-
-int
-pmemfile_mkdirat(PMEMfilepool *pfp, PMEMfile *dir, const char *path,
-		pmemfile_mode_t mode)
-{
-	struct pmemfile_vinode *at;
-	bool at_unref;
-
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!path) {
-		errno = ENOENT;
-		return -1;
-	}
-
-	if (path[0] != '/' && !dir) {
-		LOG(LUSR, "NULL dir");
-		errno = EFAULT;
-		return -1;
-	}
-
-	at = pool_get_dir_for_path(pfp, dir, path, &at_unref);
-
-	int ret = _pmemfile_mkdirat(pfp, at, path, mode);
-
-	if (at_unref)
-		vinode_cleanup(pfp, at, ret != 0);
-
-	return ret;
-}
-
-int
-pmemfile_mkdir(PMEMfilepool *pfp, const char *path, pmemfile_mode_t mode)
-{
-	return pmemfile_mkdirat(pfp, PMEMFILE_AT_CWD, path, mode);
-}
-
-/*
- * vinode_unlink_dir -- unlinks directory "vdir" from directory "vparent"
- * assuming "dirent" is used for storing this entry
- *
- * Must be called in transaction.
- */
-void
-vinode_unlink_dir(PMEMfilepool *pfp,
-		struct pmemfile_vinode *vparent,
-		struct pmemfile_dirent *dirent,
-		struct pmemfile_vinode *vdir,
-		const char *path)
-{
-	struct pmemfile_inode *iparent = vparent->inode;
-	struct pmemfile_inode *idir = vdir->inode;
-	struct pmemfile_dir *ddir = &idir->file_data.dir;
-
-	ASSERTeq(pmemobj_tx_stage(), TX_STAGE_WORK);
-
-	if (!TOID_IS_NULL(ddir->next)) {
-		LOG(LUSR, "directory %s not empty", path);
-		pmemfile_tx_abort(ENOTEMPTY);
-	}
-
-	struct pmemfile_dirent *dirdot = &ddir->dirents[0];
-	struct pmemfile_dirent *dirdotdot = &ddir->dirents[1];
-
-	ASSERTeq(strcmp(dirdot->name, "."), 0);
-	ASSERT(TOID_EQUALS(dirdot->inode, vdir->tinode));
-
-	ASSERTeq(strcmp(dirdotdot->name, ".."), 0);
-	ASSERT(TOID_EQUALS(dirdotdot->inode, vparent->tinode));
-
-	for (uint32_t i = 2; i < ddir->num_elements; ++i) {
-		struct pmemfile_dirent *d = &ddir->dirents[i];
-
-		if (!TOID_IS_NULL(d->inode)) {
-			LOG(LUSR, "directory %s not empty", path);
-			pmemfile_tx_abort(ENOTEMPTY);
-		}
-	}
-
-	pmemobj_tx_add_range_direct(dirdot, sizeof(dirdot->inode) + 1);
-	dirdot->name[0] = '\0';
-	dirdot->inode = TOID_NULL(struct pmemfile_inode);
-
-	pmemobj_tx_add_range_direct(dirdotdot,
-			sizeof(dirdotdot->inode) + 1);
-	dirdotdot->name[0] = '\0';
-	dirdotdot->inode = TOID_NULL(struct pmemfile_inode);
-
-	ASSERTeq(idir->nlink, 2);
-	TX_ADD_DIRECT(&idir->nlink);
-	idir->nlink = 0;
-
-	pmemobj_tx_add_range_direct(dirent, sizeof(dirent->inode) + 1);
-	dirent->name[0] = '\0';
-	dirent->inode = TOID_NULL(struct pmemfile_inode);
-
-	TX_ADD_DIRECT(&iparent->nlink);
-	iparent->nlink--;
-
-	struct pmemfile_time tm;
-	get_current_time(&tm);
-
-	/*
-	 * From "stat" man page:
-	 * "The field st_ctime is changed by writing or by setting inode
-	 * information (i.e., owner, group, link count, mode, etc.)."
-	 */
-	TX_SET_DIRECT(iparent, ctime, tm);
-
-	/*
-	 * From "stat" man page:
-	 * "st_mtime of a directory is changed by the creation
-	 * or deletion of files in that directory."
-	 */
-	TX_SET_DIRECT(iparent, mtime, tm);
-}
-
 /*
  * lock_parent_and_child -- resolve file with respect to parent directory and
  * lock both inodes in write mode
@@ -1404,231 +763,6 @@ race:
 	return 1;
 }
 
-int
-_pmemfile_rmdirat(PMEMfilepool *pfp, struct pmemfile_vinode *dir,
-		const char *path)
-{
-	struct pmemfile_path_info info;
-	struct pmemfile_cred cred;
-	if (cred_acquire(pfp, &cred))
-		return -1;
-
-	resolve_pathat(pfp, &cred, dir, path, &info, 0);
-
-	int error = 0;
-
-	if (info.error) {
-		error = info.error;
-		goto end;
-	}
-
-	size_t namelen = component_length(info.remaining);
-
-	/* Does not make sense, but it's specified by POSIX. */
-	if (str_compare(".", info.remaining, namelen) == 0) {
-		error = EINVAL;
-		goto end;
-	}
-
-	/*
-	 * If we managed to enter a directory, then the parent directory has
-	 * at least this entry as child.
-	 */
-	if (str_compare("..", info.remaining, namelen) == 0) {
-		error = ENOTEMPTY;
-		goto end;
-	}
-
-	if (namelen == 0) {
-		ASSERT(info.parent == pfp->root);
-		error = EBUSY;
-		goto end;
-	}
-
-
-	struct pmemfile_dirent_info dirent_info;
-	/*
-	 * lock_parent_and_child can race with another thread messing with
-	 * parent directory. Loop as long as race occurs.
-	 */
-	do {
-		error = lock_parent_and_child(pfp, &info, &dirent_info);
-	} while (error == 1);
-
-	if (error < 0) {
-		error = -error;
-		goto end;
-	}
-
-	if (!vinode_is_dir(dirent_info.vinode)) {
-		error = ENOTDIR;
-		goto vdir_end;
-	}
-
-	if (dirent_info.vinode == pfp->root) {
-		error = EBUSY;
-		goto vdir_end;
-	}
-
-	if (!_vinode_can_access(&cred, info.parent, PFILE_WANT_WRITE)) {
-		error = EACCES;
-		goto vdir_end;
-	}
-
-	TX_BEGIN_CB(pfp->pop, cb_queue, pfp) {
-		vinode_unlink_dir(pfp, info.parent, dirent_info.dirent,
-				dirent_info.vinode, path);
-
-		vinode_orphan(pfp, dirent_info.vinode);
-	} TX_ONABORT {
-		error = errno;
-	} TX_END
-
-vdir_end:
-	vinode_unlock2(dirent_info.vinode, info.parent);
-
-	if (dirent_info.vinode)
-		vinode_unref(pfp, dirent_info.vinode);
-
-end:
-	path_info_cleanup(pfp, &info);
-	cred_release(&cred);
-
-	if (error) {
-		errno = error;
-		return -1;
-	}
-
-	return 0;
-}
-
-int
-pmemfile_rmdir(PMEMfilepool *pfp, const char *path)
-{
-	struct pmemfile_vinode *at;
-	bool at_unref;
-
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!path) {
-		errno = ENOENT;
-		return -1;
-	}
-
-	at = pool_get_dir_for_path(pfp, PMEMFILE_AT_CWD, path, &at_unref);
-
-	int ret = _pmemfile_rmdirat(pfp, at, path);
-
-	if (at_unref)
-		vinode_cleanup(pfp, at, ret != 0);
-
-	return ret;
-}
-
-static int
-_pmemfile_chdir(PMEMfilepool *pfp, const struct pmemfile_cred *cred,
-		struct pmemfile_vinode *dir)
-{
-	struct inode_perms dir_perms = vinode_get_perms(dir);
-
-	if (!PMEMFILE_S_ISDIR(dir_perms.flags)) {
-		vinode_unref(pfp, dir);
-		errno = ENOTDIR;
-		return -1;
-	}
-
-	if (!can_access(cred, dir_perms, PFILE_WANT_EXECUTE)) {
-		vinode_unref(pfp, dir);
-		errno = EACCES;
-		return -1;
-	}
-
-	os_rwlock_wrlock(&pfp->cwd_rwlock);
-	struct pmemfile_vinode *old_cwd = pfp->cwd;
-	pfp->cwd = dir;
-	os_rwlock_unlock(&pfp->cwd_rwlock);
-	vinode_unref(pfp, old_cwd);
-
-	return 0;
-}
-
-int
-pmemfile_chdir(PMEMfilepool *pfp, const char *path)
-{
-	struct pmemfile_path_info info;
-	struct pmemfile_vinode *at;
-	struct pmemfile_cred cred;
-	int ret = -1;
-	int error = 0;
-	bool at_unref;
-
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!path) {
-		errno = ENOENT;
-		return -1;
-	}
-	if (cred_acquire(pfp, &cred))
-		return -1;
-
-	at = pool_get_dir_for_path(pfp, PMEMFILE_AT_CWD, path, &at_unref);
-
-	struct pmemfile_vinode *dir =
-		resolve_pathat_full(pfp, &cred, at, path, &info, 0, true);
-
-	if (info.error) {
-		error = info.error;
-		goto end;
-	}
-
-	ret = _pmemfile_chdir(pfp, &cred, dir);
-	if (ret)
-		error = errno;
-
-end:
-	path_info_cleanup(pfp, &info);
-	cred_release(&cred);
-
-	if (at_unref)
-		vinode_unref(pfp, at);
-	if (error)
-		errno = error;
-
-	return ret;
-}
-
-int
-pmemfile_fchdir(PMEMfilepool *pfp, PMEMfile *dir)
-{
-	if (!pfp) {
-		LOG(LUSR, "NULL pool");
-		errno = EFAULT;
-		return -1;
-	}
-
-	if (!dir) {
-		LOG(LUSR, "NULL dir");
-		errno = EFAULT;
-		return -1;
-	}
-
-	int ret;
-	struct pmemfile_cred cred;
-	if (cred_acquire(pfp, &cred))
-		return -1;
-	ret = _pmemfile_chdir(pfp, &cred, vinode_ref(pfp, dir->vinode));
-	cred_release(&cred);
-	return ret;
-}
-
 /*
  * pool_get_cwd -- returns current working directory
  *
@@ -1794,4 +928,10 @@ pmemfile_getcwd(PMEMfilepool *pfp, char *buf, size_t size)
 	}
 
 	return _pmemfile_get_dir_path(pfp, pool_get_cwd(pfp), buf, size);
+}
+
+size_t
+pmemfile_dir_size(TOID(struct pmemfile_dir) dir)
+{
+	return page_rounddown(pmemobj_alloc_usable_size(dir.oid));
 }
