+++ src/libpmemfile-posix/unlink.h
@@ -29,35 +29,14 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-#ifndef PMEMFILE_UTILS_H
-#define PMEMFILE_UTILS_H
+#ifndef PMEMFILE_UNLINK_H
+#define PMEMFILE_UNLINK_H
 
 #include "inode.h"
-#include "layout.h"
 
-void get_current_time(struct pmemfile_time *t);
-
-bool is_zeroed(const void *addr, size_t len);
-
-int str_compare(const char *s1, const char *s2, size_t s2n);
-bool str_contains(const char *str, size_t len, char c);
-bool more_than_1_component(const char *path);
-size_t component_length(const char *path);
-
-char *pmfi_strndup(const char *c, size_t len);
-
-#ifdef DEBUG
-const char *pmfi_path(struct pmemfile_vinode *vinode);
-#else
-static inline const char *pmfi_path(struct pmemfile_vinode *vinode)
-{
-	(void) vinode;
-	return NULL;
-}
-#endif
-
-void expand_to_full_pages(uint64_t *offset, uint64_t *length);
-void narrow_to_full_pages(uint64_t *offset, uint64_t *length);
+void vinode_unlink_file(PMEMfilepool *pfp,
+		struct pmemfile_vinode *parent,
+		struct pmemfile_dirent *dirent,
+		struct pmemfile_vinode *vinode);
 
 #endif
